feat: 타입 구분
