+++ src/types/crew/crewDetail.ts
@@ -1,0 +1,19 @@
+export interface CrewDetail {
+  crewId: number;
+  crewLevel: number;
+  crewName: string;
+  introduction: string;
+  memberCnt: number;
+  limitMemberCnt: number;
+  tags: string[];
+  category: string;
+  rule: Rule;
+  crewTotalDistance: number;
+  missionProgress: number[];
+  isCrewLeader: boolean;
+  imgFile: string | null;
+}
+export interface Rule {
+  weeklyRun: number;
+  distance: number;
+}
