+++ src/types/crew/crewList.ts
@@ -1,0 +1,25 @@
+export interface CrewListType {
+  crewId: number;
+  crewName: string;
+  crewLevel: number;
+  memberCnt: number;
+  limitMemberCnt: number;
+  approvalType: string;
+  limitRunScore: number;
+  tags: string[];
+  category: string;
+  rule: Rule;
+  introduction: string;
+  crewLeader: CrewLeader;
+}
+
+export interface CrewLeader {
+  memberId: number;
+  nickname: string;
+}
+
+export interface Rule {
+  weeklyRun: number;
+  distance: number;
+}
+
