+++ src/types/crew/crewPost.ts
@@ -1,30 +1,3 @@
-export interface CrewListType {
-  crewId: number;
-  crewName: string;
-  crewLevel: number;
-  memberCnt: number;
-  limitMemberCnt: number;
-  approvalType: string;
-  limitRunScore: number;
-  tags: string[];
-  category: string;
-  rule: Rule;
-  introduction: string;
-  crewLeader: CrewLeaderType;
-}
-
-export interface CrewLeaderType {
-  memberId: number;
-  nickname: string;
-}
-
-export interface Rule {
-  weeklyRun: number;
-  distance: number;
-}
-
-
-// ----------------------------------------------------------------
 
 export interface CrewPostList {
   id: number;
