+++ src/types/crew/crewRequest.ts
@@ -1,0 +1,19 @@
+import { Rule } from './crewList';
+import { Leader } from './crewPage';
+export interface CrewRequest {
+  crewId: number;
+  crewLevel: number;
+  crewName: string;
+  crewLeader: Leader;
+  memberCnt: number;
+  limitMemberCnt: number;
+  approvalType: string;
+  imgFile: string | null;
+  introduction: string;
+  tags: string[];
+  category: string;
+  rule: Rule;
+  crewTotalDistance: number;
+  isRequest: boolean;
+}
+
