+++ src/types/user/mission.ts
@@ -1,0 +1,9 @@
+export interface Mission {
+  missionId: number;
+  missionConent: string;
+  missonProgress: number;
+}
+
+export interface MissionList {
+  missions: Mission[];
+}
